Add validation and secure profile update
