+++ MentorHup/APPLICATION/Service/AuthServices/IMenteeAuthService.cs
@@ -8,7 +8,6 @@ namespace MentorHup.APPLICATION.Service.AuthServices
         Task<MenteeRegisterResult> RegisterAsync(MenteeRegisterRequest request);
         Task<bool> ConfirmEmailAsync(string userId, string token);
         Task<MenteeLoginResult> LoginAsync(MenteeLoginRequest request);
-        Task<bool> UpdateAsync(MenteeUpdateRequest request);
     }
 
 }
