+++ MentorHup/APPLICATION/Service/AuthServices/IMentorAuthService.cs
@@ -5,10 +5,8 @@ namespace MentorHup.APPLICATION.Service.AuthServices
     public interface IMentorAuthService
     {
         Task<MentorLoginRegistrationResult> RegisterAsync(MentorRegisterRequest request);
-        Task<UploadImageResult> UploadImageAsync(IFormFile image);
         Task<bool> ConfirmEmailAsync(string userId, string token);
         Task<MentorLoginRegistrationResult> LoginAsync(MentorLoginRequest request);
-        Task<bool> UpdateAsync(MentorUpdateRequest request);
 
     }
