+++ MentorHup/APPLICATION/Service/Mentee/IMenteeService.cs
@@ -1,0 +1,12 @@
+﻿using MentorHup.APPLICATION.Dtos.Mentee;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
+
+namespace MentorHup.APPLICATION.Service.Mentee
+{
+    public interface IMenteeService
+    {
+        Task<bool> UpdateAsync(MenteeUpdateRequest request);
+
+    }
+}
