+++ MentorHup/APPLICATION/Service/Mentor/IMentorService.cs
@@ -1,4 +1,5 @@
 ﻿using MentorHup.APPLICATION.Common;
+using MentorHup.APPLICATION.Dtos.Mentee;
 using MentorHup.APPLICATION.DTOs.Mentor;
 
 namespace MentorHup.APPLICATION.Service.Mentor;
@@ -8,5 +9,10 @@ public interface IMentorService
     Task<PageResult<MentorOverviewDto>> GetAllMentorsAsync(int pageSize ,
         int pageNumber, string? field , string? skillName , decimal? minPrice , decimal? maxPrice,
         int? Experiences);
+
+    Task<bool> UpdateAsync(MentorUpdateRequest request);
+
+    Task<UploadImageResult> UploadImageAsync(IFormFile image);
+
 
 }
