+++ MentorHup/appsettings.Development.json
@@ -7,7 +7,7 @@
   },
 
   "ConnectionStrings": {
-    "MentorHupDB": "Server = . ; Database = MentorHub4  ; Integrated Security = SSPI ; TrustServerCertificate = True;"
+    "MentorHupDB": "Server = DESKTOP-92JNGFN\\SQLEXPRESS ; Database = MentorHupDB5  ; Integrated Security = SSPI ; TrustServerCertificate = True;"
   },
 
   "JwtSettings": {
